refactor: extract `parseSFC`
